Package for ContestScoreboard
