+++ src/com/project/api/models/contestscoreboard/contest/ContestAttribute.java
