+++ src/com/project/api/models/contestscoreboard/entity/Entity.java
