+++ src/com/project/api/models/contestscoreboard/entitycharid/EntityCharIdAttribute.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.project.api.models.entitycharid;
+
+import com.project.api.models.Attribute;
+
+/**
+ *
+ * @author Kim Howel delos Reyes
+ */
+public enum EntityCharIdAttribute implements Attribute {
+    
+    ENTITY_CHAR_ID_ID;
+    
+}
