+++ src/com/project/api/models/contestscoreboard/entitylongid/EntityLongIdAttribute.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.project.api.models.entitylongid;
+
+import com.project.api.models.Attribute;
+
+/**
+ *
+ * @author Kim Howel delos Reyes
+ */
+public enum EntityLongIdAttribute implements Attribute {
+    
+    ENTITY_LONG_ID_ID;
+    
+}
