+++ src/com/project/api/models/contestscoreboard/submission/SubmissionAttribute.java
@@ -1,0 +1,19 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.project.api.models.submission;
+
+import com.project.api.models.Attribute;
+
+/**
+ *
+ * @author ieyangco
+ */
+public enum SubmissionAttribute implements Attribute {
+    
+    ATTRIBUTE_SUBMISSION_VERDICT,
+    ATTRIBUTE_SUBMISSION_TIME;
+    
+}
