+++ src/com/project/api/models/contestscoreboard/team/TeamAttribute.java
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.project.api.models.team;
+
+import com.project.api.models.Attribute;
+
+/**
+ *
+ * @author Kim Howel delos Reyes
+ */
+public enum TeamAttribute implements Attribute {
+    
+    ATTRIBUTE_TEAM_ID,
+    ATTRIBUTE_TEAM_SCHOOL_NAME,
+    ATTRIBUTE_TEAM_TEAM_NAME,
+    ATTRIBUTE_TEAM_SUBMISSIONS,
+    ATTRIBUTE_TEAM_PROBLEM_SOLVED;
+    
+}
