+++ src/com/project/api/models/contestscoreboard/user/User.java
